requirements.txt
pandas
numpy
yfinance
ccxt
fredapi
cvxpy
PyPortfolioOpt
arch
statsmodels
scikit-learn
xgboost
backtrader
plotly
matplotlib
streamlit
pyyaml
pydantic